+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## vNEXT
+- Add configuration for deployments on custom test networks. (#37)
+- Verify that voucher type exists when adding or removing eligible asset. (#36)
 - Deploy contracts through generic factory. (#35)
 - Define admin addresses for deployment (#34)
 - Mark assembly blocks as memory-safe and update deal price calculation syntax. (#33)